Refactor expense model to transaction, update related references, and add user details model

+++ Server/app/Http/Controllers/BudgetController.php
@@ -14,23 +14,12 @@ use Illuminate\Support\Facades\Gate;
 class BudgetController extends Controller
 {
 
-    /**
-     * Display a listing of the resource.
-     */
     public function index()
     {
-        $budgets = Budget::where('user_id', Auth::id())->get();;
-        return response()->json($budgets);
+        $budgets = Budget::where('user_id', Auth::id())->get();
+        return response()->json(['budgets' => $budgets], Response::HTTP_OK);
     }
 
-    /**
-     * Show the form for creating a new resource.
-     */
-    public function create() {}
-
-    /**
-     * Store a newly created resource in storage.
-     */
     public function store(Request $request)
     {
         $request->validate([
@@ -53,15 +42,14 @@ class BudgetController extends Controller
         $budget->category_id = $request->category_id;
         $budget->custom_period = $request->custom_period;
         $budget->save();
-        return response()->json($budget);
+
+        return response()->json(['message' => 'Budget created successfully', 'budget:' => $budget], Response::HTTP_CREATED);
     }
 
-    /**
-     * Display the specified resource.
-     */
     public function show(string $id)
     {
         $budget = Budget::find($id);
+
         if (!$budget) {
             return response()->json(['message' => 'Budget not found'], Response::HTTP_NOT_FOUND);
         }
@@ -70,41 +58,61 @@ class BudgetController extends Controller
         if ($authResponse->denied()) {
             return response()->json(['message' => $authResponse->message()], Response::HTTP_FORBIDDEN);
         }
-        return response()->json($budget);
+
+        return response()->json(['budget' => $budget], Response::HTTP_OK);
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     */
-    public function edit(string $id)
-    {
-        //
-    }
-
-    /**
-     * Update the specified resource in storage.
-     */
     public function update(Request $request, string $id)
     {
         $budget = Budget::find($id);
+
+        if (!$budget) {
+            return response()->json(['message' => 'Budget not found'], Response::HTTP_NOT_FOUND);
+        }
+
+        $authResponse = Gate::inspect('update', $budget);
+        if ($authResponse->denied()) {
+            return response()->json(['message' => $authResponse->message()], Response::HTTP_FORBIDDEN);
+        }
+
+        $request->validate([
+            'name' => 'string|max:250',
+            'limit' => 'numeric|min:0|max:999999',
+            'period' => 'string|in:daily,weekly,monthly,custom',
+            'start_date' => 'date|date_format:Y-m-d|after_or_equal:today',
+            'category_id' => 'exists:categories,id',
+            'custom_period' => 'integer|min:1|max:365|nullable'
+        ]);
+
+        $user = Auth::user();
+
+        $budget->user_id = $user->id;
         $budget->name = $request->name;
+        $budget->limit = $request->limit;
+        $budget->period = $request->period;
+        $budget->start_date = $request->start_date;
+        $budget->category_id = $request->category_id;
+        $budget->custom_period = $request->custom_period;
         $budget->save();
-        return response()->json($budget, Response::HTTP_OK);
+
+        return response()->json(['message' => 'Budget updated successfully', 'budget:' => $budget], Response::HTTP_OK);
     }
 
-    /**
-     * Remove the specified resource from storage.
-     */
     public function destroy(string $id)
     {
         $budget = Budget::find($id);
-        $budget->delete();
-        return response()->json($budget);
-    }
 
-    public function getByUserId(string $id)
-    {
-        $budgets = User::find($id)->budgets;
-        return response()->json($budgets);
+        if (!$budget) {
+            return response()->json(['message' => 'Budget not found'], Response::HTTP_NOT_FOUND);
+        }
+
+        $authResponse = Gate::inspect('delete', $budget);
+        if ($authResponse->denied()) {
+            return response()->json(['message' => $authResponse->message()], Response::HTTP_FORBIDDEN);
+        }
+
+        $budget->delete();
+
+        return response()->json(['message' => 'Budget deleted successfully'], Response::HTTP_OK);
     }
 }
